django/requirements.txt
asgiref==3.5.2
CD==1.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.3
djangorestframework==3.14.0
idna==3.4
Markdown==3.4.1
pytz==2022.6
requests==2.28.1
sqlparse==0.4.3
stripe==4.2.0
urllib3==1.26.12
